Move category support into it's own trait
